Module Input

OCaml bindings for libinput. This library is used by e.g. Wayland compositors to support keyboards and mice.

Contexts

module Context : sig ... end

Handles for accessing libinput.

module Interface : sig ... end

Callbacks for opening device files.

module Udev : sig ... end

Minimal bindings to libudev to avoid an external dependency for simple uses of libinput.

Devices

module Device : sig ... end

Input devices.

module Device_group : sig ... end

A group of kernel devices making up one physical device.

module Config_status : sig ... end

Status codes returned when applying configuration settings.

module Seat : sig ... end

A collection of input devices (e.g. mouse and keyboard) to be used together by one person.

Events

module Event : sig ... end

Events (e.g. keypresses) reported to the application from libinput.

module Timestamp : sig ... end

Event timestamps, with microsecond precision.

module Keycode : sig ... end

Keyboard key code.

Tablets

module Tablet_tool : sig ... end

A tool being used by a device with the Device.Capability.t.Tablet_tool capability.

module Mode_group : sig ... end

A mode on a tablet pad is a virtual grouping of functionality.