Module Context.Path

A context that requires the caller to add and remove devices explicitly.

val create : Interface.t -> [> `Path ] t

create interface returns a fresh, empty libinput context.

Call destroy to release the context when finished (GC will not free it).

val add_device : [ `Path ] t -> string -> Device.t option

val remove_device : Device.t -> unit