Module Context.Udev

A context that uses udev to find available devices.

val create : Interface.t -> Udev.t -> [> `Udev ] t

This context is inactive until assigned a seat ID with assign_seat.

Call destroy to release the context when finished (GC will not free it).

val assign_seat : [ `Udev ] t -> string -> unit

assign_seat t id assigns a seat to this libinput context.

New devices or the removal of existing devices will appear as events during dispatch.

This call succeeds even if no input devices are currently available on this seat, or if devices are available but fail to open (with Interface.t.open_restricted).

Devices that do not have the minimum capabilities to be recognized as pointer, keyboard or touch device are ignored. Such devices and those that failed to open are ignored until the next call to resume.

This function may only be called once per context.