Module Event.Keyboard

Events from keyboard devices.

type ty = [

| `Keyboard_key of [ `Keyboard_key ] t

]

val get_key : [ `Keyboard_key ] t -> Keycode.t

val get_key_state : [ `Keyboard_key ] t -> [ `Released | `Pressed ]

val get_time : [ `Keyboard_key ] t -> Timestamp.t

val get_seat_key_count : [ `Keyboard_key ] t -> int

get_seat_key_count t is the total number of instances of this key being pressed across all devices

This is typically either 0 (for a release) or 1 (for a press). It's useful if you have e.g. two keyboards and press Shift on both, then release on one of them, to know that Shift is still pressed somewhere.